+++ src/collective/documentgenerator/helper/archetypes.py
@@ -61,29 +61,6 @@ class ATDocumentGenerationHelperView(DocumentGenerationHelperView):
 
         return display
 
-    def display_text(self, field_name):
-        text = self.get_value(field_name)
-        return self.portal.portal_transforms.convert('web_intelligent_plain_text_to_html', text).getData()
-
-    def display_html(self, field_name):
-        if not self.appy_renderer:
-            return ''
-
-        html_text = self.get_value(field_name)
-        display = self.appy_renderer.renderXhtml(html_text)
-        return display
-
-    def display_list(self, field_name, separator=', '):
-        values = self.get_value(field_name)
-        display = separator.join(values)
-
-        return display
-
-    def list(self, field_name):
-        values = self.get_value(field_name)
-
-        return values
-
 
 class ATDisplayProxyObject(DisplayProxyObject):
     """